+++ src/aa_utilities/graphics/__init__.py
@@ -1,3 +1,4 @@
 
 from ._graphics import *
 from ._interactive_selection import InteractiveSelection
+from ._repel_text import repel_text